requirements.txt
tqdm
mysql-connector-python
requests
redis
pandas